Protect match invariant more obviously
* The "retry when encountering an exception upon instantiating a match
between a participant and themselves" mechanism worked sufficiently,
but it was not very obvious about the invariant it protected: namely
a `ParticipantsMatch` requires the donor and the recipient to be
different `Participant` instances. Therefore -- even though it is
actually superfluous -- I added an extra guard in the calling class to
shine light on what is being done and prevent future changes from
breaking the matching algorithm (e.g. when the class itself might not
protect invariants anymore).