Suppress validation on add/remove participants
* Since this is not the "final" submission of the form, but merely a
modification of a yet unfinished/unconfirmed event it seemed cruel to
burden the user with input validation all the time. So I suppressed the
validation until the final "submit" call and to the backend (will
follow). This was done by migrating the add/remove `button` elements to
`input` elements with `formnovalidate` enabled.
* This also allowed simplifying the structure of the UI tests dealing
  with add/remove. I decided to combine the tests for adding and
  removing since they shared a big chunk of code anyway and are
  relatively slow.